requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
bleach==3.2.1
bokeh==2.2.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
cycler==0.10.0
dash==1.16.3
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
dateparser==0.7.6
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
hyperlink==20.0.1
idna==2.10
impact==0.5.9
incremental==17.5.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
pprintpp==0.4.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
PyHamcrest==2.0.2
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.10.23
requests==2.24.0
retrying==1.3.3
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.5
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zope.interface==5.1.2
